added polymorphism module

+++ src/exercise/Calculator.java
@@ -20,7 +20,7 @@ public class Calculator {
         return num1 * num2;
     }
 
-    public float divideNumbers(int num1, int num2){
+    public float divideNumbers(float num1, float num2){
         return num1/num2;
     }
